Move StatBlock to common
