+++ client/Commands/Prompts/PrivacyPolicyPrompt.tsx
@@ -1,6 +1,6 @@
 import * as React from "react";
+import { StatBlock } from "../../../common/StatBlock";
 import { Button } from "../../Components/Button";
-import { StatBlock } from "../../StatBlock/StatBlock";
 import { Metrics } from "../../Utility/Metrics";
 import { Store } from "../../Utility/Store";
 import { Prompt } from "./Prompt";
