+++ server/user.ts
@@ -1,7 +1,5 @@
 import mongo = require("mongodb");
 
-import { StatBlock } from "../client/StatBlock/StatBlock";
-
 export interface User {
     _id: mongo.ObjectID;
     settings: any;
